Refactored to allow history usage:
